extract menus reducer into flux/parts/menus
